+++ packages/extension-base/src/background/KoniTypes.ts
@@ -138,6 +138,14 @@ export interface ResultResolver {
   accounts: string[];
 }
 
+// Switch current network auth
+
+export interface RequestSwitchCurrentNetworkAuthorization {
+  url: string;
+  networkKey: string;
+  authSwitchNetworkType: AccountAuthType
+}
+
 /// Staking subscribe
 
 export enum StakingType {
@@ -191,7 +199,24 @@ export interface PriceJson {
   currencyData: CurrencyJson,
   exchangeRateMap: Record<string, ExchangeRateJSON>,
   priceMap: Record<string, number>,
-  price24hMap: Record<string, number>
+  price24hMap: Record<string, number>,
+  priceCoinGeckoSupported: string[],
+  lastUpdatedMap: Record<string, Date>
+}
+
+export interface HistoryTokenPriceJSON {
+  history: PriceChartPoint[];
+}
+
+export interface ResponseSubscribeCurrentTokenPrice {
+  id: string;
+  price: CurrentTokenPrice;
+}
+
+export interface CurrentTokenPrice {
+  value: number;
+  value24h: number;
+  time: number;
 }
 
 export interface ExchangeRateJSON {
@@ -415,6 +440,13 @@ export type CurrencyType = 'USD'
 | 'RUB'
 | 'VND'
 
+export type PriceChartTimeframe = '1D' | '1W' | '1M' | '3M' | 'YTD' | '1Y' | 'ALL';
+
+export interface PriceChartPoint {
+  time: number;
+  value: number;
+}
+
 export type LanguageOptionType = {
   text: string;
   value: LanguageType;
@@ -463,6 +495,8 @@ export type RequestChangeShowZeroBalance = { show: boolean };
 export type RequestChangeLanguage = { language: LanguageType };
 
 export type RequestChangePriceCurrency = { currency: CurrencyType }
+
+export type RequestGetHistoryTokenPriceData = { priceId: string, timeframe: PriceChartTimeframe };
 
 export type RequestChangeShowBalance = { enable: boolean };
 
@@ -1132,6 +1166,7 @@ export interface TonSignatureRequest extends TonSignRequest {
 export interface CardanoSignatureRequest extends CardanoSignRequest {
   id: string;
   errors?: ErrorValidation[];
+  currentAddress: string;
   payload: unknown
 }
 
@@ -1140,6 +1175,14 @@ export interface EvmSendTransactionRequest extends TransactionConfig, EvmSignReq
   parseData: EvmTransactionData;
   isToContract: boolean;
   errors?: ErrorValidation[]
+}
+
+export interface SubmitApiRequest extends EvmSignRequest {
+  id: string;
+  type: string;
+  payload: unknown;
+  errors?: ErrorValidation[];
+  processId?: string;
 }
 
 // Cardano Request Dapp Input
@@ -1273,7 +1316,8 @@ export interface ConfirmationDefinitions {
   evmSignatureRequest: [ConfirmationsQueueItem<EvmSignatureRequest>, ConfirmationResult<string>],
   evmSendTransactionRequest: [ConfirmationsQueueItem<EvmSendTransactionRequest>, ConfirmationResult<string>]
   evmWatchTransactionRequest: [ConfirmationsQueueItem<EvmWatchTransactionRequest>, ConfirmationResult<string>],
-  errorConnectNetwork: [ConfirmationsQueueItem<ErrorNetworkConnection>, ConfirmationResult<null>]
+  errorConnectNetwork: [ConfirmationsQueueItem<ErrorNetworkConnection>, ConfirmationResult<null>],
+  submitApiRequest: [ConfirmationsQueueItem<SubmitApiRequest>, ConfirmationResult<string>]
 }
 
 export interface ConfirmationDefinitionsTon {
@@ -2151,6 +2195,9 @@ export interface KoniRequestSignatures {
   // Price, balance, crowdloan functions
   'pri(price.getPrice)': [RequestPrice, PriceJson];
   'pri(price.getSubscription)': [RequestSubscribePrice, PriceJson, PriceJson];
+  'pri(price.getHistory)': [RequestGetHistoryTokenPriceData, HistoryTokenPriceJSON];
+  'pri(price.checkCoinGeckoPriceSupport)': [string, boolean];
+  'pri(price.subscribeCurrentTokenPrice)': [string, ResponseSubscribeCurrentTokenPrice, CurrentTokenPrice];
   'pri(balance.getBalance)': [RequestBalance, BalanceJson];
   'pri(balance.getSubscription)': [RequestSubscribeBalance, BalanceJson, BalanceJson];
   'pri(crowdloan.getCrowdloan)': [RequestCrowdloan, CrowdloanJson];
@@ -2389,6 +2436,7 @@ export interface KoniRequestSignatures {
   'cardano(account.get.address)': [null, string[]];
   'cardano(account.get.balance)': [null, Cbor];
   'cardano(account.get.change.address)': [null, string];
+  'cardano(account.get.reward.address)': [null, string[]];
   'cardano(account.get.utxos)': [RequestCardanoGetUtxos, Cbor[] | null];
   'cardano(account.get.collateral)': [RequestCardanoGetCollateral, Cbor[] | null];
   'cardano(network.get.current)': [null, number];