+++ packages/extension-base/src/services/chain-service/index.ts
@@ -1523,6 +1523,11 @@ export class ChainService {
             deprecated = true;
             break;
           }
+
+          if (defaultChainAsset.slug === storedAssetInfo.slug) {
+            duplicated = true;
+            break;
+          }
         }
 
         if (!duplicated && !deprecated) {